+++ src/main/java/com/vanndeth/phoneshopmgt/service/implement/BrandServiceImplement.java
@@ -33,13 +33,13 @@ public class BrandServiceImplement implements BrandService {
 	}
 
 	@Override
-	public Brand getById(Integer id) {
+	public Brand getById(Long id) {
 		return brandRepository.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("Brand", id));
 	}
 
 	@Override
-	public Brand update(Integer id, Brand newBrand) {
+	public Brand update(Long id, Brand newBrand) {
 		Brand brand = getById(id);
 		brand.setName(newBrand.getName());
 		return brandRepository.save(brand);
